requirements.txt
# PostgreSQL driver + ORM
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.30

# Excel / CSV parsing
openpyxl>=3.1.2          # .xlsx support
xlrd>=2.0.1              # legacy .xls support (BIFF8)
pandas>=2.2.0

# Environment & CLI
python-dotenv>=1.0.1
click>=8.1.7

# Logging
colorlog>=6.8.2

# Testing (dev)
pytest>=8.2.0
pytest-mock>=3.14.0
